Added ability to listen to a specific IP/CIDR
It’s now possible to have only a given IP/CIDR redirected to the
listener. This is useful if you only want a specific IP to perform the
egress bust (i.e. during a penetration test), but more importantly it
means that the server can continue to be used normally (for other
tasks), as any other client connecting to the server’s services will
work as per normal. This allows Egress Buster to be used on a
multi-purpose server.

Also cleaned up the way the firewall rule is removed when Egress Buster
is stopped. It now cleanly removes the added rule instead of flushing
the entire iptables. Useful if the server relies on a rule set,
especially in a multi-purpose server scenario.